GoWeb clase 3 .env + persistiendo datos en otro archivo

+++ 8_goweb3/ClaseTM/ProyectoEstructura/cmd/server/handler/productoController.go
@@ -1,6 +1,8 @@
 package handler
 
 import (
+	"fmt"
+	"os"
 	"strconv"
 
 	producto "github.com/extmatperez/meli_bootcamp2/tree/scerca_nahuel/8_goweb3/ClaseTM/ProyectoEstructura/internal/producto"
@@ -22,18 +24,34 @@ type ProductoController struct {
 	service producto.ServiceProducto
 }
 
+func validarToken(ctx *gin.Context) (bool, int, error) {
+
+	tok := ctx.GetHeader("token")
+	if os.Getenv("TOKEN") == tok {
+		return true, 200, nil
+	} else {
+		return false, 403, fmt.Errorf("Token incorrecto: %s", tok)
+	}
+}
+
 func NewProductoController(ser producto.ServiceProducto) *ProductoController {
 	return &ProductoController{service: ser}
 }
 
 func (prodController *ProductoController) GetAll() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
-		personas, err := prodController.service.GetAll()
+		validado, status, err := validarToken(ctx)
+		fmt.Println(validado, err, status)
+		if validado {
+			personas, err := prodController.service.GetAll()
 
-		if err != nil {
-			ctx.String(400, "Hubo un error %v", err)
+			if err != nil {
+				ctx.String(400, "Hubo un error %v", err)
+			} else {
+				ctx.JSON(200, personas)
+			}
 		} else {
-			ctx.JSON(200, personas)
+			ctx.JSON(status, err.Error())
 		}
 	}
 }
@@ -41,48 +59,60 @@ func (prodController *ProductoController) GetAll() gin.HandlerFunc {
 func (prodController *ProductoController) Store() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
 
-		var produc request
+		validado, status, err := validarToken(ctx)
+		if validado {
+			var produc request
 
-		err := ctx.ShouldBindJSON(&produc)
+			err := ctx.ShouldBindJSON(&produc)
 
-		if err != nil {
-			ctx.String(400, "Hubo un error al querer cargar una persona %v", err)
-		} else {
-			response, err := prodController.service.Store(produc.Nombre, produc.Color, produc.Precio, produc.Stock, produc.Codigo, produc.Publicado, produc.FechaCreacion)
 			if err != nil {
-				ctx.String(400, "No se pudo cargar la persona %v", err)
+				ctx.String(400, "Hubo un error al querer cargar una persona %v", err)
 			} else {
-				ctx.JSON(200, response)
+				response, err := prodController.service.Store(produc.Nombre, produc.Color, produc.Precio, produc.Stock, produc.Codigo, produc.Publicado, produc.FechaCreacion)
+				if err != nil {
+					ctx.String(400, "No se pudo cargar la persona %v", err)
+				} else {
+					ctx.JSON(200, response)
+				}
 			}
-		}
+		} else {
+			ctx.JSON(status, err.Error())
 
+		}
 	}
 }
 
 func (prodController *ProductoController) Update() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
-		var produc request
 
-		id, err := strconv.Atoi(ctx.Param("id"))
+		validado, status, err := validarToken(ctx)
 
-		if err == nil {
-			err := ctx.ShouldBindJSON(&produc)
+		if validado {
+			var produc request
+
+			id, err := strconv.Atoi(ctx.Param("id"))
 
 			if err == nil {
-				prod, err := prodController.service.Update(id, produc.Nombre, produc.Color, produc.Precio, produc.Stock, produc.Codigo, produc.Publicado, produc.FechaCreacion)
+				err := ctx.ShouldBindJSON(&produc)
 
 				if err == nil {
-					ctx.JSON(200, prod)
+					prod, err := prodController.service.Update(id, produc.Nombre, produc.Color, produc.Precio, produc.Stock, produc.Codigo, produc.Publicado, produc.FechaCreacion)
+
+					if err == nil {
+						ctx.JSON(200, prod)
+					} else {
+						ctx.JSON(404, gin.H{
+							"message": err.Error(),
+						})
+					}
 				} else {
-					ctx.JSON(404, gin.H{
-						"message": err.Error(),
-					})
+					ctx.String(400, "Error con los datos del body")
 				}
 			} else {
-				ctx.String(400, "Error con los datos del body")
+				ctx.String(400, "Error con el id")
 			}
 		} else {
-			ctx.String(400, "Error con el id")
+			ctx.JSON(status, err.Error())
 		}
 
 	}
@@ -90,47 +120,62 @@ func (prodController *ProductoController) Update() gin.HandlerFunc {
 
 func (prodController *ProductoController) Detele() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
-		id, err := strconv.Atoi(ctx.Param("id"))
 
-		if err == nil {
-			err := prodController.service.Delete(id)
+		validado, status, err := validarToken(ctx)
+
+		if validado {
+			id, err := strconv.Atoi(ctx.Param("id"))
+
 			if err == nil {
-				ctx.String(200, "Producto %d borrado exitosamente", id)
+				err := prodController.service.Delete(id)
+				if err == nil {
+					ctx.String(200, "Producto %d borrado exitosamente", id)
+				} else {
+					ctx.JSON(400, gin.H{
+						"message": err.Error(),
+					})
+				}
 			} else {
-				ctx.JSON(400, gin.H{
-					"message": err.Error(),
-				})
+				ctx.String(400, "Error con el id")
 			}
 		} else {
-			ctx.String(400, "Error con el id")
+			ctx.JSON(status, err.Error())
 		}
+
 	}
 }
 
 func (prodController *ProductoController) UpdateNameAndPrice() gin.HandlerFunc {
 	return func(ctx *gin.Context) {
-		id, err := strconv.Atoi(ctx.Param("id"))
-		var prod request
-		errorCast := ctx.ShouldBindJSON(&prod)
-		if err == nil {
-			if errorCast == nil {
-				productoMod, errorMod := prodController.service.UpdateNameAndPrice(id, prod.Nombre, prod.Precio)
-				if errorMod == nil {
-					ctx.JSON(200, gin.H{
-						"message":  "Campos Modificados correctamente",
-						"producto": productoMod,
-					})
-				} else {
-					ctx.JSON(200, gin.H{
-						"message": errorMod.Error(),
-					})
-				}
-			} else {
-				ctx.String(200, "Hubo un error con body, asegurese de pasar bien los campos nombre y precio")
-			}
+		validado, status, err := validarToken(ctx)
 
+		if validado {
+			id, err := strconv.Atoi(ctx.Param("id"))
+			var prod request
+			errorCast := ctx.ShouldBindJSON(&prod)
+			if err == nil {
+				if errorCast == nil {
+					productoMod, errorMod := prodController.service.UpdateNameAndPrice(id, prod.Nombre, prod.Precio)
+					if errorMod == nil {
+						ctx.JSON(200, gin.H{
+							"message":  "Campos Modificados correctamente",
+							"producto": productoMod,
+						})
+					} else {
+						ctx.JSON(200, gin.H{
+							"message": errorMod.Error(),
+						})
+					}
+				} else {
+					ctx.String(200, "Hubo un error con body, asegurese de pasar bien los campos nombre y precio")
+				}
+
+			} else {
+				ctx.String(200, "Hubo un error con el id")
+			}
 		} else {
-			ctx.String(200, "Hubo un error con el id")
+			ctx.JSON(status, err.Error())
 		}
+
 	}
 }
